requirements.txt
# Core dependencies
Pillow>=10.0.0  # For image processing and metadata extraction